requirements.txt
streamlit
openai
python-dotenv
langchain
langchain_huggingface
langchain_community
langchain_openai
faiss-cpu
PyPDF2
python-docx
python-pptx
pandas
openpyxl
rank_bm25 
beautifulsoup4








